fix(ui): matchCard alignment, clip event type, CSS vars
