+++ packages/web/src/App.css
@@ -74,6 +74,11 @@
   --sidebar-accent-foreground: oklch(0.205 0 0);
   --sidebar-border: oklch(0.922 0 0);
   --sidebar-ring: oklch(0.708 0 0);
+  --victory: oklch(0.53 0.18 145);
+  --defeat: oklch(0.577 0.245 27.325);
+  --info: oklch(0.55 0.15 255);
+  --gold: oklch(0.75 0.17 85);
+  --warning: oklch(0.7 0.19 60);
 }
 
 .dark {
@@ -108,6 +113,11 @@
   --sidebar-accent-foreground: oklch(0.985 0 0);
   --sidebar-border: oklch(1 0 0 / 10%);
   --sidebar-ring: oklch(0.556 0 0);
+  --victory: oklch(0.53 0.18 145);
+  --defeat: oklch(0.577 0.245 27.325);
+  --info: oklch(0.55 0.15 255);
+  --gold: oklch(0.75 0.17 85);
+  --warning: oklch(0.7 0.19 60);
 }
 
 @layer base {
